infrastructure/auto-scaling/app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
docker==6.1.3
paramiko==3.3.1
pyyaml==6.0.1
apscheduler==3.10.4
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
prometheus-client==0.19.0